requirements.txt
pillow==11.3.0
PyMuPDF==1.26.1
